added sNode_test.go

+++ keyI.go
@@ -1,5 +1,5 @@
 package ctries_go
 
 type KeyI interface {
-	Hashcode()	int64
+	Hashcode() uint32
 }
